requirements.txt
Django==5.2.8
Pillow==11.1.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
